Support passing end_of_stream to header callbacks
Do this in a backwardly-compatible way by creating two new ABI
functions:

on_request_headers_ext
on_response_headers_ext

Each one takes an additional Word parameter at the end which
carries the value of the "end_of_stream" flag from Envoy.
The proxy_wasm_cpp_host code will first check for the existence
of the "ext" version of a function and fall back to the original
version if it is not found.

Signed-off-by: Gregory Brail <gregbrail@google.com>