Pass end_of_stream to header callbacks.
This adds the same end_of_stream flag to the on_request_headers
and on_response_headers callbacks that Envoy makes for non-WASM filters.
This way it is possible to write a filter that can operate on the
request or response body but still be resilient if no body is present.

Signed-off-by: Gregory Brail <gregbrail@google.com>